# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Personal Finance & Investment Tracker API",
    "description": "API for tracking personal finances, investments (Gold and Mutual Funds), and receiving AI-powered investment recommendations.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "description": "Creates a new user account with encrypted email and hashed password.",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "email": "user@example.com",
                "password": "securepassword123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                },
                "example": {
                  "responseCode": 201,
                  "responseStatus": "SUCCESS",
                  "responseMessage": "User registered successfully",
                  "responseDetails": {
                    "id": "1",
                    "email": "user@example.com",
                    "created_at": "2026-01-02T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cashflow": {
      "get": {
        "tags": ["Cashflow"],
        "summary": "Get cashflow history",
        "description": "Retrieves all monthly cashflow records for the authenticated user.",
        "operationId": "getCashflowHistory",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Cashflow history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowHistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Cashflow"],
        "summary": "Create or update cashflow",
        "description": "Creates a new cashflow record or updates an existing one for the specified month. Automatically calculates total_expense and net_cashflow.",
        "operationId": "saveCashflow",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashflowInput"
              },
              "example": {
                "month": "2026-01",
                "income": 10000000,
                "expense_rent": 2000000,
                "expense_living": 3000000,
                "expense_other": 1000000
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cashflow saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analytics/recommendation": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get AI investment recommendation",
        "description": "Retrieves AI-powered investment allocation recommendation based on user's financial history. Returns 403 if AI recommendation is disabled in user settings.",
        "operationId": "getInvestmentRecommendation",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Investment recommendation generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationResponse"
                },
                "example": {
                  "responseCode": 200,
                  "responseStatus": "SUCCESS",
                  "responseMessage": "Investment recommendation generated successfully",
                  "responseDetails": {
                    "gold_percentage": 40,
                    "mutual_fund_percentage": 60,
                    "investable_amount": 4000000,
                    "reasoning": "Berdasarkan analisis keuangan Anda, kami merekomendasikan alokasi 40% untuk Emas dan 60% untuk Reksa Dana. Cashflow Anda stabil dengan rata-rata surplus Rp 4.000.000 per bulan.",
                    "risk_profile": "moderate",
                    "should_invest": true,
                    "warnings": []
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "AI recommendation disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "responseCode": 403,
                  "responseStatus": "ERROR",
                  "responseMessage": "AI recommendation is disabled. Enable it in settings to receive investment recommendations.",
                  "responseDetails": null
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get user settings",
        "description": "Retrieves the current user's settings including AI recommendation preference.",
        "operationId": "getUserSettings",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Settings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                },
                "example": {
                  "responseCode": 200,
                  "responseStatus": "SUCCESS",
                  "responseMessage": "Settings retrieved successfully",
                  "responseDetails": {
                    "ai_recommendation_enabled": true
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications/send-monthly": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Trigger monthly notifications",
        "description": "Triggers monthly email notifications for all users. Sends reminder emails to users without cashflow data for the current month, and summary emails to users with data. This endpoint is designed to be called by a cron job scheduler.",
        "operationId": "sendMonthlyNotifications",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Monthly notifications processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResultResponse"
                },
                "example": {
                  "responseCode": 200,
                  "responseStatus": "SUCCESS",
                  "responseMessage": "Monthly notifications processed: 5 sent, 0 failed",
                  "responseDetails": {
                    "sent": 5,
                    "failed": 0,
                    "total": 5,
                    "details": [
                      { "userId": "1", "type": "SUMMARY", "success": true },
                      { "userId": "2", "type": "REMINDER", "success": true }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (invalid cron secret)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from NextAuth.js authentication"
      }
    },
    "schemas": {
      "APIResponse": {
        "type": "object",
        "required": ["responseCode", "responseStatus", "responseMessage", "responseDetails"],
        "properties": {
          "responseCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "responseStatus": {
            "type": "string",
            "enum": ["SUCCESS", "ERROR"],
            "description": "Operation status"
          },
          "responseMessage": {
            "type": "string",
            "description": "Human-readable message"
          },
          "responseDetails": {
            "description": "Operation-specific data or null"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "User password (minimum 8 characters)"
          }
        }
      },
...
      "RegisterResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseDetails": {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "email": { "type": "string" },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        ]
      },
      "CashflowInput": {
        "type": "object",
        "required": ["month", "income", "expense_rent", "expense_living", "expense_other"],
        "properties": {
          "month": {
            "type": "string",
            "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
            "description": "Month in YYYY-MM format"
          },
          "income": {
            "type": "number",
            "minimum": 0,
            "description": "Monthly income"
          },
          "expense_rent": {
            "type": "number",
            "minimum": 0,
            "description": "Rent expense"
          },
          "expense_living": {
            "type": "number",
            "minimum": 0,
            "description": "Living expense"
          },
          "expense_other": {
            "type": "number",
            "minimum": 0,
            "description": "Other expenses"
          }
        }
      },
      "CashflowRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "user_id": { "type": "string" },
          "month": { "type": "string" },
          "income": { "type": "number" },
          "expense_rent": { "type": "number" },
          "expense_living": { "type": "number" },
          "expense_other": { "type": "number" },
          "total_expense": { "type": "number" },
          "net_cashflow": { "type": "number" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "CashflowResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseDetails": { "$ref": "#/components/schemas/CashflowRecord" }
            }
          }
        ]
      },
      "CashflowHistoryResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseDetails": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/CashflowRecord" }
              }
            }
          }
        ]
      },
...
      "ValidationErrorResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseCode": { "type": "integer", "example": 400 },
              "responseStatus": { "type": "string", "example": "ERROR" },
              "responseDetails": {
                "type": "object",
                "properties": {
                  "errors": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                }
              }
            }
          }
        ]
      },
      "UnauthorizedResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseCode": { "type": "integer", "example": 401 },
              "responseStatus": { "type": "string", "example": "ERROR" },
              "responseMessage": { "type": "string", "example": "Unauthorized" },
              "responseDetails": { "type": "null" }
            }
          }
        ]
      },
...
      "ErrorResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseStatus": { "type": "string", "example": "ERROR" },
              "responseDetails": { "type": "null" }
            }
          }
        ]
      },
      "ServerErrorResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseCode": { "type": "integer", "example": 500 },
              "responseStatus": { "type": "string", "example": "ERROR" },
              "responseMessage": { "type": "string", "example": "Internal server error" },
              "responseDetails": { "type": "null" }
            }
          }
        ]
      },
...
      "AllocationRecommendation": {
        "type": "object",
        "properties": {
          "gold_percentage": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Recommended percentage allocation for Gold (0-100)"
          },
          "mutual_fund_percentage": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "Recommended percentage allocation for Mutual Fund (0-100)"
          },
          "investable_amount": {
            "type": "number",
            "description": "Suggested amount to invest this month"
          },
          "reasoning": {
            "type": "string",
            "description": "Explanation in Bahasa Indonesia for the recommendation"
          },
          "risk_profile": {
            "type": "string",
            "enum": ["conservative", "moderate", "aggressive"],
            "description": "User's determined risk profile"
          },
          "should_invest": {
            "type": "boolean",
            "description": "Whether user should invest (false if should focus on emergency fund)"
          },
          "warnings": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Any warnings or caveats"
          }
        }
      },
      "RecommendationResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseDetails": { "$ref": "#/components/schemas/AllocationRecommendation" }
            }
          }
        ]
      },
      "UserSettings": {
        "type": "object",
        "properties": {
          "ai_recommendation_enabled": {
            "type": "boolean",
            "description": "Whether AI investment recommendations are enabled"
          }
        }
      },
      "SettingsResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseDetails": { "$ref": "#/components/schemas/UserSettings" }
            }
          }
        ]
      },
...
      "NotificationResult": {
        "type": "object",
        "properties": {
          "sent": { "type": "integer", "description": "Number of notifications sent successfully" },
          "failed": { "type": "integer", "description": "Number of notifications that failed" },
          "total": { "type": "integer", "description": "Total notifications processed" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "userId": { "type": "string" },
                "type": { "type": "string", "enum": ["REMINDER", "SUMMARY"] },
                "success": { "type": "boolean" }
              }
            }
          }
        }
      },
      "NotificationResultResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/APIResponse" },
          {
            "type": "object",
            "properties": {
              "responseDetails": { "$ref": "#/components/schemas/NotificationResult" }
            }
          }
        ]
      }
    }
  }
}